Little extra info on the accept_keys
I've tried every one of these kind of plugins over the years and I think this is the best. It just behaves exactly as I would want. Thanks!

Just adding a little extra info about the `accept_keys` here as I think it is an incredibly smart piece of detail that was lost on me initially and I fear maybe other users too. I modded my keys as a Colemak user and now pounce is even better!